+++ src/modules/private/issuer/businessApplication/components/Documentation.js
@@ -19,6 +19,7 @@ export default class Documentation extends Component {
       businessAppRemoveFiles,
       getBusinessTypeCondtion,
       getPersonalGuaranteeCondition,
+      formReadOnlyMode,
     } = this.props.businessAppStore;
     const { fields } = BUSINESS_DOC_FRM;
     const statementFileList = getBusinessTypeCondtion ? ['bankStatements', 'leaseAgreementsOrLOIs'] : ['leaseAgreementsOrLOIs'];
@@ -55,6 +56,7 @@ export default class Documentation extends Component {
                   statementFileList.map(field => (
                     <Grid.Column>
                       <DropZone
+                        disabled={formReadOnlyMode}
                         multiple
                         key={field}
                         name={field}
@@ -90,6 +92,7 @@ export default class Documentation extends Component {
                   taxFileList.map(field => (
                     <Grid.Column>
                       <DropZone
+                        disabled={formReadOnlyMode}
                         multiple
                         key={field}
                         name={field}
@@ -109,6 +112,7 @@ export default class Documentation extends Component {
               subHeader="NextSeed will require it. (Note that if you have existing debt with liens attached, a second lien will be accepted.)"
             >
               <FormRadioGroup
+                disabled={formReadOnlyMode}
                 fielddata={fields.blanketLien}
                 name="blanketLien"
                 changed={businessDocChange}
@@ -120,6 +124,7 @@ export default class Documentation extends Component {
               subHeader="(This is not a requirement, but a personal guarantee can positively impact the terms provided.)"
             >
               <FormRadioGroup
+                disabled={formReadOnlyMode}
                 fielddata={fields.personalGuarantee}
                 name="personalGuarantee"
                 changed={businessDocChange}
@@ -132,6 +137,7 @@ export default class Documentation extends Component {
                     Personal Guarantee Form along with any supporting documentation
                   </p>
                   <DropZone
+                    disabled={formReadOnlyMode}
                     multiple
                     name="personalGuaranteeForm"
                     fielddata={fields.personalGuaranteeForm}